requirements.txt
langchain
langchain-google-genai
google-ai-generativelanguage==0.6.18
python-dotenv
streamlit
